+++ src/static.js
@@ -1,0 +1,6 @@
+export default class Static {
+  constructor({ folderPath = "", headers = {} } = {}) {
+    this.folderPath = folderPath;
+    this.headers = headers;
+  }
+}